pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-screener"
version = "0.1.0"
description = "Financial analysis and screening system for Kubernetes"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Financial Screener Team"}
]
keywords = ["finance", "stocks", "screening", "kubernetes", "raspberry-pi"]

dependencies = [
    # Core data validation and settings
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",

    # Database drivers
    "asyncpg>=0.28.0,<0.30.0",
    "psycopg2-binary>=2.9.0,<3.0.0",

    # Logging
    "structlog>=23.0.0,<24.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",

    # Code quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",

    # Pre-commit hooks
    "pre-commit>=3.4.0",
]

data-collector = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
]

analyzer = [
    "polars>=0.19.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]

api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "strawberry-graphql>=0.209.0",
]

[tool.setuptools]
packages = ["financial_screener"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
financial_screener = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=financial_screener --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/financial_screener"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
